+++ Sources/scs/specticus.swift
@@ -1,13 +1,31 @@
 import ArgumentParser
+import Foundation
+
+// MARK: - Entry
+
+@main
+enum ScsMain {
+    static func main() {
+        // Custom chrome for --version / --help (#135); then ArgumentParser.
+        if Brand.isRootVersionInvocation(arguments: CommandLine.arguments) {
+            Brand.printWordmarkIfTTY()
+            print(Brand.version)
+            return
+        }
+        if Brand.isHelpInvocation(arguments: CommandLine.arguments) {
+            Brand.printWordmarkIfTTY()
+        }
+        Specticus.main()
+    }
+}
 
 // MARK: - CLI Root
 
-@main
 struct Specticus: ParsableCommand {
     static let configuration = CommandConfiguration(
         commandName: "scs",
         abstract: "A fast, beautiful documentation generator for specifications as code.",
-        version: "0.1.0",
+        version: Brand.version,
         subcommands: [
             Build.self,
             Init.self,